refactor: Renamed dataset folder to mrfi_dataset to improve clarity. Updated examples and readme to suit the changes
