+++ experiments/poisson_sample.py
@@ -1,5 +1,5 @@
 '''Compare speed of poisson_sample'''
-from dataset.imagenet import make_testloader
+from mrfi_dataset.imagenet import make_testloader
 from mrfi import MRFI, EasyConfig
 from mrfi.experiment import BER_Acc_experiment, logspace_density, Acc_golden
 from torchvision.models import resnet18
