+++ experiments/resnet18_BER_Acc.py
@@ -1,5 +1,5 @@
 
-from dataset.imagenet import make_testloader
+from mrfi_dataset.imagenet import make_testloader
 from mrfi import MRFI, EasyConfig
 from mrfi.experiment import BER_Acc_experiment, logspace_density, Acc_golden, benchmark_range
 import matplotlib.pyplot as plt
