+++ experiments/resnet18_weight.py
@@ -1,5 +1,5 @@
 
-from dataset.imagenet import make_testloader
+from mrfi_dataset.imagenet import make_testloader
 from mrfi import MRFI, EasyConfig
 from mrfi.experiment import BER_Acc_experiment, logspace_density, Acc_golden, get_weight_info
 from torchvision.models import resnet18
