+++ mrfi_dataset/lenet_cifar10.pth
