+++ readme.md
@@ -47,7 +47,7 @@ The following code perform a quantized random integer bit flip injection on LeNe
 and find the relation between bit error rate (BER) and classification accuracy.
 
 ```python title="LeNet default fault injection"
-from dataset.lenet_cifar import make_testloader, LeNet
+from mrfi_dataset.lenet_cifar import make_testloader, LeNet
 from mrfi import MRFI, EasyConfig
 from mrfi.experiment import Acc_experiment, Acc_golden, BER_Acc_experiment
